project/requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
Django==4.1.5
django-cors-headers==3.14.0
djangorestframework==3.14.0
mysql-client==0.0.1
mysqlclient==2.1.1
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
